Added exception handling for CRUD methods (exception package)

+++ src/main/java/com/unlimitedparts/demo/controller/BrandController.java
@@ -1,5 +1,6 @@
 package com.unlimitedparts.demo.controller;
 import com.unlimitedparts.demo.domain.Brand;
+import com.unlimitedparts.demo.exception.ApiRequestException;
 import com.unlimitedparts.demo.service.DTO.BrandDTO;
 import com.unlimitedparts.demo.service.request.CreateBrandRequest;
 import com.unlimitedparts.demo.service.BrandService;
@@ -22,11 +23,11 @@ public class BrandController {
 
     @PostMapping
     public ResponseEntity<String> addBrand(@RequestBody CreateBrandRequest brandRequest){
-        if (brandRequest != null){
+        if (brandRequest.getBrandName() != null){
             brandService.addBrand(brandRequest);
             return ResponseEntity.ok("Successfully added brand.");
         }
-        return ResponseEntity.badRequest().body("Could not add brand.");
+        throw new ApiRequestException("Could not add car-brand to the database.");
     }
 
     @GetMapping("/{brandId}")
@@ -35,7 +36,7 @@ public class BrandController {
         if (dto != null){
             return ResponseEntity.ok(dto);
         }
-        return ResponseEntity.notFound().build();
+        throw new ApiRequestException("Could not return car-brand from the database.");
     }
 
     @DeleteMapping("/{brandId}")
@@ -43,11 +44,12 @@ public class BrandController {
         Optional<Brand> brandOptional = brandService.getBrandById(brandId);
         if (brandOptional.isPresent()){
             Brand brand = brandOptional.get();
+            if(!brand.getCars().isEmpty()) throw new ApiRequestException("Could not delete car-brand from the database because it is still linked to the cars.");
             brand.getCars().clear();
             brandService.deleteBrandById(brandId);
             return ResponseEntity.ok("Brand successfully deleted.");
         }
-        return ResponseEntity.notFound().build();
+        throw new ApiRequestException("Could not delete car-brand from the database.");
     }
 
 }
